Refactor encryption algorithm
Changes:

1. Add final to variables.
2. Change directory for factory class.
