+++ src/test/java/com/zysn/passwordmanager/model/security/algorithm/encryption/impl/AESTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 
 import com.zysn.passwordmanager.model.security.algorithm.config.impl.AlgorithmConfig;
 import com.zysn.passwordmanager.model.security.algorithm.encryption.api.EncryptionAlgorithm;
-import com.zysn.passwordmanager.model.security.algorithm.encryption.patterns.EncryptionAlgorithmFactory;
+import com.zysn.passwordmanager.model.security.algorithm.encryption.factory.EncryptionAlgorithmFactory;
 import com.zysn.passwordmanager.model.utils.encoding.EncodingUtils;
 
 public class AESTest {
@@ -42,7 +42,8 @@ public class AESTest {
         this.algorithmConfig = new AlgorithmConfig(algorithmName, algorithmType);
         this.algorithmConfig.setSalt(iv);
 
-        this.algorithmConfig.addNewParameter("variant", "AES/GCM/NoPadding");
+        this.algorithmConfig.addNewParameter("mode", "GCM");
+        this.algorithmConfig.addNewParameter("padding", "NoPadding");
     }
 
     @Test
@@ -77,4 +78,5 @@ public class AESTest {
         assertArrayEquals(expectedDecryptedData, actualDecryptedData,
                 "The algorithm didn't create the right decrypted data.");
     }
+
 }
